+++ src/directives.js
@@ -57,10 +57,14 @@ angular.module('oi.select')
                     multiplePlaceholder = multiplePlaceholderFn(scope),
                     elementOptions      = optionsFn(scope.$parent) || {},
                     options             = angular.extend({cleanModel: elementOptions.newItem === 'prompt'}, oiSelect.options, elementOptions),
-                    editItem            = options.editItem === true || options.editItem === 'correct' ? 'oiSelectEditItem' : options.editItem,
-                    editItemCorrect     = options.editItem === 'correct',
-                    editItemFn          = editItem ? $injector.get(editItem) : function() {return ''},
-                    removeItemFn        = $parse(options.removeItemFn);
+                    editItem            = options.editItem,
+                    editItemIsCorrected = editItem === 'correct';
+
+                if (editItem === true || editItem === 'correct') {
+                    editItem = 'oiSelectEditItem';
+                }
+                var editItemFn   = editItem ? $injector.get(editItem) : angular.noop,
+                    removeItemFn = $parse(options.removeItemFn);
 
                 match = options.searchFilter.split(':');
                 var searchFilter = $filter(match[0]),
@@ -83,7 +87,7 @@ angular.module('oi.select')
                     };
                 }
 
-                if (options.cleanModel && (!editItem || editItemCorrect)) {
+                if (options.cleanModel && (!editItem || editItemIsCorrected)) {
                     element.addClass('cleanMode');
                 }
 
@@ -133,11 +137,16 @@ angular.module('oi.select')
                     }
                 });
 
-                scope.$parent.$watch(attrs.ngModel, function(value) {
+                scope.$parent.$watch(attrs.ngModel, function(value, oldValue) {
                     var output = value instanceof Array ? value : value ? [value]: [],
                         promise = $q.when(output);
 
                     modifyPlaceholder();
+
+                    if (value !== oldValue) {
+                        editItemIsCorrected = false;
+                        element.removeClass('cleanMode');
+                    }
 
                     if (!multiple) {
                         restoreInput();
@@ -268,13 +277,8 @@ angular.module('oi.select')
                                 }
                             }
 
-                            if (!editItemCorrect && (multiple || !scope.backspaceFocus)) {
-                                scope.query = editItemFn(removedItem, lastQuery, getLabel);
-                            }
-
-                            if (editItem) {
-                                editItemCorrect = false;
-                                element.removeClass('cleanMode');
+                            if (multiple || !scope.backspaceFocus) {
+                                scope.query = editItemFn(removedItem, lastQuery, getLabel, editItemIsCorrected);
                             }
 
                             if (multiple && options.closeList) {
@@ -429,7 +433,7 @@ angular.module('oi.select')
                     }
 
                     if (scope.isOpen && options.closeList && (event.target.nodeName !== 'INPUT' || !scope.query.length)) { //do not reset if you are editing the query
-                        resetMatches({query: options.editItem && !editItemCorrect});
+                        resetMatches({query: options.editItem && !editItemIsCorrected});
                         scope.$evalAsync();
                     } else {
                         getMatches(scope.query);